Portal Gem Fixes
Do not consume a Lifestone or Portal Recall gem if there is no linked destination.
Do still punish the player for being a dork by running the animation and invoking the timer.
Also, if an item is used on a target and the result is not successful (e.g. try to use a portal tie on a portal you cannot use, such as an outpost portal), then return the appropriate error message for that target. Previously, it was attempting to return the wrong (result) error, rather than the target item (resultTarget) message. For some reason I don't immediately see, it also left the client in waiting mode, which required a client restart to resolve.